Migrate pulses to dashboard subscriptions
couple things to note:
- just to call out, this is removing old style pulses (where you just
choose individual questions to send) NOT alerts. This is why the initial
creation is `INSERT INTO report_dashboard ...FROM pulse p WHERE
p.dashboard_id IS NULL AND alert_condition IS NULL;` AKA, not alert, and
where it doesn't point at a dashboard.
- the UI of pulses ensures that they have a name, but some on stats
don't, perhaps from an older time. This explains the coalesce on pulse
name when creating new dashboards
- the schema allows for a single pulse to have multiple channels. In
practice this is a pulse has both an email and a slack
component. However, dashboard subscriptions enforce a single pulse to a
single channel. The UI will not show the second pulse channel of a pulse
and there's no way to edit it because of this. This is why each pulse
channel is the source of new pulses, so that one pulse can become
two. Requires a bit more book keeping but no real problem
- just a surprising bit: the pulse_channel_recipient table is not the
only place for recipients. The details map has the slack channel for
slack type messages. We also put in that details map email addresses
that are typed in free form. If you select a person in the UI to receive
the email it goes in the recipients channel. eg,

```
pulse_channel.details        | array_to_string(array_agg(u.email), ', ')
{"emails":["test@test.com"]} | foo@metabase.com, bar@metabase.com
```